+++ sources/filesystem/gitpythonfs/gitpythonfs/__init__.py
@@ -1,3 +1,3 @@
-from .core import GitPythonFileSystem, register_implementation_in_fsspec
+from gitpythonfs.core import GitPythonFileSystem, register_implementation_in_fsspec
 
 register_implementation_in_fsspec()
